samples/python/hosted-agents/langgraph/react-agent-with-foundry-tools/requirements.txt
azure-ai-agentserver-langgraph==1.0.0b9